refactor(database): emit UserRecoveryCodes SETs without coalesce
Drop the patch/coalesce helpers so each typed update writes one SET.
Keep RequireNonEmptyRecoveryCodes on codes updates. Use uint8 for
increment Delta so negatives are a compile error.

+++ internal/domain/user_recovery_codes_update.go
@@ -37,7 +37,7 @@ type UserRecoveryCodesLastSuccessfulCheckUpdate struct {
 func (*UserRecoveryCodesLastSuccessfulCheckUpdate) userRecoveryCodesUpdate() {}
 
 type UserRecoveryCodesIncrementFailedAttemptsUpdate struct {
-	Delta int16
+	Delta uint8
 }
 
 func (*UserRecoveryCodesIncrementFailedAttemptsUpdate) userRecoveryCodesUpdate() {}
